+++ kturner/src/main/java/playground/kturner/guiceAdvanced2019/Main.java
@@ -1,4 +1,4 @@
-package playground.kturner.guiceInMatsimWithoutControler;
+package playground.kturner.guiceAdvanced2019;
 
 import com.google.inject.*;
